Added TestWiring, renamed OpModes directories so they would be together.
